refactor: profile section
