+++ src/sections/user/components/profile-user-info.tsx
@@ -2,7 +2,7 @@ import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
 import BadgeVerified from "@src/components/user-item/BadgeVerified.tsx";
 import {FC} from "react";
-import {ProfileHeaderProps} from "@src/sections/user/profile-header.tsx";
+import {ProfileHeaderProps} from "@src/sections/user/components/profile-header.tsx";
 import { Address } from "viem";
 
 const ProfileUserInfo: FC<ProfileHeaderProps> = ({profile}) => {
