+++ src/sections/user/index.ts
@@ -1,0 +1,1 @@
+export { default as UserProfileView } from '@src/sections/user/views/user-profile-view.tsx';
