--- src/sections/user/view/index.ts
@@ -1,1 +1,0 @@
-export { default as UserProfileView } from './user-profile-view';
